requirements.txt
google-generativeai==0.8.3
google-api-python-client==2.134.0
openai>=1.52.0
python-dotenv>=1.0.0
requests>=2.31.0
asyncio>=3.4.3
aiohttp>=3.9.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
supabase>=2.0.0
boto3>=1.28.0
gunicorn>=21.2.0
